requirements.txt
numpy~=1.19.5
tensorflow~=2.4.1
scipy~=1.6.2
Pillow~=7.0.0